requirements.txt
streamlit==1.26.0
torch==2.3.0
torchvision==0.18.0
numpy
pillow
streamlit-drawable-canvas
